refactor: optimize size of axis/ and label/
