+++ packages/vrender-components/src/label/arc.ts
@@ -20,6 +20,7 @@ import {
   computeQuadrant
 } from './util';
 import type { ComponentOptions } from '../interface';
+import { registerLabelComponent } from './data-label-register';
 
 export class ArcInfo {
   key!: string;
@@ -970,3 +971,7 @@ export class ArcLabel extends LabelBase<ArcLabelAttrs> {
     return this.computeLayoutRadius(width ? width : 0, height ? height : 0) * outerRadiusRatio;
   }
 }
+
+export const registerArcDataLabel = () => {
+  registerLabelComponent('arc', ArcLabel);
+};
