+++ packages/vrender-components/src/label/data-label-register.ts
@@ -1,0 +1,9 @@
+const labelComponentMap = {};
+
+export const registerLabelComponent = (type: string, LabelClass: any) => {
+  labelComponentMap[type] = LabelClass;
+};
+
+export const getLabelComponent = (type: string) => {
+  return labelComponentMap[type];
+};
