+++ packages/vrender-components/src/label/index.ts
@@ -4,7 +4,5 @@ export * from './rect';
 export * from './line';
 export * from './base';
 export * from './arc';
-export * from './area';
-export * from './line-data';
 
 export * from './dataLabel';
